Split off qoreVscPkg functions to a separate file
